[linalg] StencilVector class: manually flag outdated ghost regions.
The previous choice of relying on __setitem__() has proven to be
dangerous, because such a method is not always called collectively by
all processes in the MPI communicator, and therefore the '_sync' flag
may have different values on different MPI ranks.

As a consequence, it may happen that not all processes in the MPI
communicator make a call to 'update_ghost_regions()', which must be
called collectively in order to run correctly; ultimately this leads
to an MPI deadlock.

After this modification the user should take care of setting the
property 'ghost_regions_in_sync' to False when appropriate. Failure
to do so may result in incorrect results, but never in an MPI deadlock.