Remove obsolete code generation for periodic case.
This is not needed anymore, because the code that is generated for the
non-periodic case also works for a periodic domain. Moreover, the old
special code was inefficient, not well tested, and it was not working
if the spline degree was not the same along all directions.

To verify that the periodic case is treated properly, a unit test is added:
2D Poisson's equation on a unit circle, with periodic BCs along the theta
direction, homogeneous Dirichlet BC at r=1, and different spline degrees
along r and theta.