Fix bug in function 'compute_normal_vector' of api/ast/utilities:
The normal vector is identified with a dual basis vector of the
coordinate transformation (reversed if needed). Hence it is computed
as a row of the inverse of the Jacobian matrix of the mapping. This
procedure is simple and general, and it applies to domains with any
number of dimensions.

TODO: verify if a normalization is required.