Increase tolerance in test_histopolation_exact
This test in psydac/fem/tests/test_spline_histopolation.py uses a random polynomial profile, and from time to time the resulting error is larger than the tolerance. This results in unexpected failures in our test suite. We have now increased the relative and absolute error tolerances from 1e-12 to 1e-11 in order to ameliorate this problem.