Always reassemble if free arguments are present
If a free argument is a FemField, there is no simple way to know if this
has changed since the last assembly event. So it is pointless to try and
determine if reassembly is necessary based on the values of the free
arguments. Instead it is safer to assume that every free argument always
changes between successive assembly calls.