Fix `__div__` -> `__truediv__` (#126)
Update the division operator to Python 3 (assuming that Python 2 support is obsolete/removed already),
enabling us to do e.g. `M /= a` where `M` is a `Matrix` or `Vector` object, and `a` is something that we can invert and have `M *= a` implemented, e.g. `a` could be a scalar:

* Change `__div__` with `__truediv__`, and `__idiv__` with `__itruediv__`;
* Add simple tests for `StencilMatrix` and `StencilVector` concerning some of their operators.

Note: we do not implement `__floordiv__` here (if that should be needed at all).