Fix distributed/parallel spline assembly when using DeRham complexes (#132)
Fix a bug in relation to the `FemAssemblyGrid` when `pad` does not equal `degree`. Previously, there seemed to be some discrepancies on multiple processes (except for the first one) at least on a periodic grid which caused a segmentation fault or an out of bounds error in the kernel, depending on if pyccel was used or not. Now, the spans are elevated on all processes, if `pad==degree+1`. (i.e. the padding parameter is used instead of the degree parameter).

Commit summary:

* Make DeRham work in parallel;
* Take pad parameter directly from `SplineSpace`;
* (Sort of) fix `local_element_start` and `local_element_end` in corner cases, where `end < degree`;
* Include small unit tests;
* Add parallel unit tests to files `test_api_feec_1d.py` and `test_api_feec_2d.py`, with check on L2 error norm.