Update igakit's URL. Build it w/ same Numpy version as Numba. Add pyproject.toml (#207)
Igakit (not on PyPI) was just moved to GitHub, hence we are updating its URL in our `setup.py` file. This solves #206.

Because of the incompatibility between the Numpy build versions for Numba and Igakit, we now install Igakit without a pip isolated environment in order to use the same Numpy version that was previously installed by Numba. To this end we now list Igakit in a new file, `requirements_extra.txt`, which is dedicated to non-isolated builds.

In addition, we add the new file `pyproject.toml` in accordance to PEP 518.