Improve names of generated functions (#200)
All assembly functions used to have the generic name 'assembly', which caused problems on macOS when using the MPI compilers (probably because of caching). We now provide names that are more informative, and that contain a unique `tag` identifier:

- `assemble_matrix_{tag}` for bilinear forms,
- `assemble_vector_{tag}` for linear forms,
- `assemble_scalar_{tag}` for functional forms.

In addition, we make sure that the accelerated functions for matrix-vector product and matrix transposition also have a `tag` identifier, which is unique across all processes. Finally, we run unit tests on macOS (fixes #201). 

Co-authored-by: Noureddine Hamid <nhamid@student.1337.ma>