Enforce contiguity in `bsplines.py` (#224)
In `fem/grid.py`, line 69-70, we reverse the order of arrays before giving them to `quadrature_grid`. Those views are not contiguous and thus need to be copied in a contiguous way before giving them to the pyccelised function in `bsplines_pyccel.py`. This wasn't done before and thus caused issue #223. 

To avoid further issues of the same kind, in `bsplines.py` we have replaced most of the calls to `np.asarray` with calls to `np.ascontiguousarray`.