Update ghost regions before assembling a form (#281)
When assembling a discrete linear/bilinear form (or functional) which takes a `FemField` as parameter, make sure that the spline coefficients have up-to-date ghost regions. Failure to do so may yield wrong results.

Add a unit test showing how the error could show up. The test failed prior to the bug fix.