Do not decompose domain along user-specified dimensions (#286)
Add new flag `mpi_dims_mask` to `discretize_domain`, which is
successively passed to:
- `ddm.cart.DomainDecomposition`
- `ddm.partition.compute_dims`
- `ddm.partition.compute_dims_general`

Function `compute_dims_general` is the ones responsible to create the
Cartesian grid of MPI processes. It receives `mpi_dims_mask` which is
either `None` or a list of `bool`. In the latter case each element is
set to `True` if the dimension is to be used in the domain decomposition
(=default for each dimension). If `mpi_dim_mask[i] == False`, the i-th
dimension is not decomposed.

Two unit tests for 2d and 3d have been added.

---------

Co-authored-by: Yaman Güçlü <yaman.guclu@gmail.com>