Replace quad_order with nquads (#289)
Rename all instances of `quad_order` with `nquads` as intended by PR
#174.

Note: `nquads` is the number of quadrature points (along each direction)
for a Gaussian quadrature rule. In the future this parameter will be
mandatory when creating a `DiscreteFunctional`, a `DiscreteLinearForm`,
or a `DiscreteBilinearForm`. Instead, passing this argument to the
constructor of `TensorFemSpace` will be forbidden.

Further, allow `TensorFemSpace` objects to store `d` dictionaries of 1D
`FemAssemblyGrid` objects, with `d` being the number of dimensions of
the domain. Every time the user asks for the `d` grids by calling
`TensorFemSpace.get_quadrature_grids(*nquads)`, the grid along dimension
`i` is retrieved from the `i`-th dictionary if the key `nquads[i]` is
present, otherwise a new grid is computed and stored in the dictionary.

---------

Co-authored-by: Yaman Güçlü <yaman.guclu@gmail.com>