Do not access '_name' attribute of object with no name (#328)
When collecting a common subexpression from a long expression in the
AST, a temporary variable is created whose name is obtained from the
`_name` attribute of the first SymPy object in a list. In certain
unfortunate cases such an object does not have the `_name` attribute and
an `AttributeError` is raised.

The problem is solved by cycling through each element in the
aforementioned list, and extracting `_name` from the first object with
such an attribute. This is done via the new `get_name` function. Fixes
#327.