Add `recycle` flag to linear solvers' constructors (#334)
Add a `recycle` flag to the `__init__` function of the subclasses of
`InverseLinearOperator` defined in `linalg/solvers`.
By default `False`. If `True`, updates the initial guess `x0` to be a
copy of the previous solution, without creating a temporary vector.

This is especially beneficial when the `InverseLinearOperator`
object is a term in a `LinearOperator` expression (which could be,
e.g., a `ComposedLinearOperator` or a `BlockLinearOperator`),
and therefore its `dot` method cannot be called manually with the
correct value of `x0`.