Moved _check_options() and transpose() from the solver classes to the base class `InverseLinearOperator`.
All solvers call now `super().__init__` which takes the base arguments for any iterative solver.
Additional arguments must be checked in the individual solvers.

The base class now has the setter method for `linop`, which allows to change tha matrix A of the solver.
Moreover, there is the abstract property `solver` which is the name string of the solver.