Fix parallel bug in `simplify` method of `SumLinearOperator` (#367)
Maintain the original order of the addends in the method `simplify` of class `SumLinearOperator`. This is achieved by using `dict` (which preserves the insertion order as part of the language specification as of Python 3.7) in place of `set` (which does not keep an order).

As a consequence of the deterministic ordering, every process in a parallel run now has the same identical addends. Earlier this was not always the case, which resulted in random errors.

A minimal reproducible example is not included because the bug appears to be system- and configuration-dependent.