Clean up linear solvers, add `StencilDiagonalMatrix` and `diagonal()`, require pc to be `LinearOperator` (#360)
Reduce code duplication in the linear solvers by adding the method
`__init__` to the common base class `InverseLinearOperator`. Add new
class `StencilDiagonalMatrix` to be used in combination with the new
method `diagonal` of `StencilMatrix` and `BlockLinearOperator` -- this
fixes #369. Require all preconditioners to be `LinearOperator` objects
-- this fixes #343.

Detailed list of changes
-----------------------

All solvers are sub-classes of `InverseLinearOperator`. As such, they
now call now `super().__init__` which takes the base arguments for any
iterative solver. The method `_check_options` is moved to the base class
and checks the common parameters `x0`, `tol`, `maxiter` and `verbose`.
Additional arguments must be checked in the individual solvers.

The methods `get_options`, `set_options` and `transpose` are also moved
to the base class. In particular `get_options` has been improved as it
accepts now a specific option name. The property `options` has been
removed because unsafe (it gave direct access to the underlying
dictionary without performing any check). The method `set_options` must
be used instead.

The base class now has a setter method for `linop`, which allows to
change the matrix A of the solver.

Moreover, changed the handling of preconditioners in
`InverseLinearOperator`s. The option `pc` must now take as value a
`LinearOperator`, which is the preconditioner. If `pc=None`, an
`IdentityOperator` of the domain space is created automatically.

The new class `StencilDiagonalMatrix` has been added. This represents a
linear operator acting between two `StencilVectorSpace`s, and which
corresponds to a diagonal matrix. This linear operator is completely
local.

The method `diagonal` has been added to the classes
`StencilDiagonalMatrix` and `BlockLinearOperator` (for which it returns
an object of the same class), and to the class `StencilMatrix` (for
which it returns a `StencilDiagonalMatrix` object). As well as an `out`
optional argument, this method accepts the `inverse` boolean flag which
is useful for creating a Jacobi preconditioner.
    
The new solver `PBiConjugateGradientStabilized` (preconditioned
BiCGSTAB) has been added from Struphy (does not work for complex).
    
`test_solvers.py` has been updated, as well as several API unit tests.

Notes
-----
When passing a preconditioner to `DiscreteEquation.set_solver()`, this
also has to be a `LinearOperator` now. This means that in the case of a
Jacobi preconditioner one needs first to assemble the linear system,
then extract its diagonal, and finally pass the new `LinearOperator` as
the `pc` solver option.

---------

Co-authored-by: Yaman Güçlü <yaman.guclu@gmail.com>