Fix several errors in `cell_index_p` (#392)
In `cell_index_p` we look for cells containing points. If a point is
outside the domain exactly at `tol` distance from the bounds, then the
search loop will be stuck in infinite recursion. One way to fix this is
to lower the tolerance when checking if the points are within the domain
(before actually entering the loop).

Also fix the `status` value returned by the cell search in
`cell_index_p`, so that it is:

  * 0 if all points are in domain
  * -1 if some points are outside domain as expected by calling
`cell_index` (was 1 before)
  * -2 if search can't decide

In the third case the previous code was actually caught in an infinite
loop. To avoid this behaviour, a max number of iterations is added to
`cell_index_p`.

A unit test was added.