Fixed isinstance() error raised by sphinx-gallery
The call-order of sphinx extensions seems to have some side-effect
that conflicts with `sphinx-gallery`. I'm not able to pinpoint
which extension/s cause this; as a workaround, I moved sphinx_gallery
to the top of extension list. This reordering of sphinx extensions
seem to magically fix the cryptic `isinstance()` issue.