Pass `nquads` when discretizing integral expressions, not spaces (#403)
Allow the user to choose the number of quadrature points by passing the
keyword-only argument `nquads`:
- When calling `discretize` on integral expressions (`BilinearForm`,
`LinearForm`, `Functional`);
- When creating a `GlobalProjector` object which requires histopolation;
- When calling the method `integral` of `TensorFemSpace`.

`nquads` should be a list of integers where each entry represents the
number of quadrature points (per element) along a different direction.
(A Gauss-Legendre quadrature rule is used on each element of the
discrete logical domain.)

In the case of a multipatch domain, the same values in `nquads` are used
for all the patches. If the user passes a single integer instead of a
list, the same number of quadrature points is used in all directions. If
the user does not pass `nquads` at all, a default value is computed
along each direction `i` by checking the highest spline degree
`max_degree[i]` and then setting `nquads[i] = max_degree[i] + 1`.

`nquads` can no longer be passed to a call to `discretize` on a
`FemSpace`. Accordingly, the property `nquads` and the method
`quad_grids` have been removed from the class `TensorFemSpace`. Instead,
the property `nquads` has been added to the classes
`DiscreteBilinearForm`, `DiscreteLinearForm`, and `DiscreteFunctional`.

By attaching `nquads` to the discrete forms rather than the discrete
spaces, we allow for more flexibility in the discretization step.
Internally this is achieved by storing multiple 1D `FemAssemblyGrid`
objects in a single `TensorFemSpace` object. These grids are computed on
demand based on the required values of `nquads`, and stored in the
object so they can be reused later.

Additional changes:

- In class`TensorFemSpace` remove method `quad_grids` and rename method
`get_quadrature_grids` as `get_assembly_grids`;
- When the default values do not suffice, pass the correct `nquads` in
the various unit tests in order to obtain exactly the same results as
before;
- Make the function `gauss_legendre` in `psydac.utilities.quadratures`
accept the number of quadrature points `m` explicitly, rather than the
obscure `ordergl`. Further, fix order of abscissas (previously
reversed);
- Fix command to obtain `HDF5_DIR` on macOS;
- Use SymPDE version 0.18.3 (this gives control over the linearity
checks on linear and bilinear forms).

Fixes #332.