Improve conversion to PETSc (#408)
Implement a new conversion to PETSc which reduces the communication in
the assembly to almost none. The converted `PETSc.Vec` and `PETSc.Mat`
objects locally contain the same coefficients as the local
`StencilVector` or `StencilMatrix`/`BlockLinearOperator`. This is done
with an index conversion function which maps the Psydac natural
multi-index to the PETSc global index.

A pyccelizable kernel for computing the I,J,V format of PETSc has been
added.

The conversion is not implemented for:
- `BlockVector` with at least a block that is not a `StencilVector`.
- `BlockLinearOperator` with at least a block that is not a
`StencilMatrix`.
- Matrix-free `LinearOperator`s.

---------

Co-authored-by: Yaman Güçlü <yaman.guclu@gmail.com>