Avoid installing h5py from cache (#430)
Add the flag `--no-cache-dir` to `pip` in order to guarantee
that `h5py` is always built during installation. See:

  - https://stackoverflow.com/questions/45594707/what-is-pips-no-cache-dir-good-for;
  - https://docs.h5py.org/en/stable/build.html#custom-installation.

Re-using a prebuilt `h5py` (*i.e.* cached `h5py`) could become an issue
when `h5py` dependencies upgrade. For example, If people have a cached
version of the h5py library which was compiled using NumPy < 2.0 and try
to install Psydac in a clean Python environment, they end up with a broken
version. This is because the cached version of h5py tries to call C functions
from NumPy 2.0 which are not ABI compatible.

---------

Co-authored-by: Yaman Güçlü <yaman.guclu@gmail.com>