Improve FEM API (names, properties, constructors...) & plot_field() (#468)
Improve the FEM API to make it more user-friendly and compatible between
single-patch and multi-patch APIs (fix #459 and fix #460):

* Rename abstract property `vector_space` of `FemSpace` as `coeff_space`
* Rename `ProductFemSpace` as `MultipatchFemSpace`
* Do not use class constructors as factories (previously
`ProductFemSpace()` could return a `VectorFemSpace` object)
* Provide more specific helper properties for FEM spaces, such as
- `is_vector_valued` and `is_multipatch` (also remove a method
`is_scalar` which was implemented in _some_ FEM spaces but not used)
- `patch_spaces`, `component_spaces` and `axis_spaces` which always
return tuples of Fem spaces (or an error)
* Provide more specific helper properties for FEM fields, such as
- `component_fields` and `patch_fields` and `axis_fields` which always
return tuples of FEM fields (or an error)
* Improve the `VectorFemSpace` class (in particular its constructor) to
better match the `FemSpace` abstract interface.

Since these changes are also motivated by a better compatibility between
single-patch and multi-patch APIs (see issue #331), this PR also
modifies the `plot_field` function (previously in
`psydac.feec.multipatch.plotting_utilities`), which is:

* made compatible with single patch and multi patch fields
* renamed as `plot_field_2d` since it only handles 2d fields
* tested (not the content but the runs) in single and multipatch
configurations for scalar and vector-valued FEM fields
* moved with its module `psydac.feec.multipatch.plotting_utilities` to
`psydac.fem.plotting_utilities` since it is not specific to FEEC or
multipatch.

---------

Co-authored-by: Yaman Güçlü <yaman.guclu@gmail.com>