Add diagonal() to KroneckerStencilMatrix (#537)
We add a `diagonal(self, inverse=False, sqrt=False, out=None)` method to
the class `KroneckerStencilMatrix`.
Additionally, we fix the `diagonal` method of `StencilDiagonalMatrix`,
both making it run in general, and also allowing for the kwarg `sqrt`
which had been missing until now.

This method comes in handy for the computation of the yet to be
implemented LST preconditioners.
For the computation of these, we compute 1D mass matrices which are the
building blocks of 2D or 3D logical mass matrices. Additionally, we
compute the diagonal of said 2D or 3D logical mass matrix.
Thus, instead of having to assemble the entire 2D or 3D mass matrix, we
can build them as `KroneckerStencilMatrices`, which is much cheaper, and
then obtain the diagonal.

This feature reduces the runtime of a yet to be implemented LST
preconditioner test from 2 minutes down to 20 seconds.

A very fast test for this new method is included in this PR.